[FIX] SecurityConfig -> GlobalSecurityConfig , 인증 관련 로직 안정화
