+++ src/main/java/com/kidmily/algoga_server/admin/settings/AdminJwtAuthenticationFilter.java
@@ -20,46 +20,36 @@ public class AdminJwtAuthenticationFilter extends OncePerRequestFilter {
 
     private final AdminJwtProvider adminJwtProvider; // 어드민 전용 프로바이더 사용
 
-    // 어드민 API가 아닌 요청(유저 요청 등)은 이 필터를 무시하고 패스하도록 설정
-    @Override
-    protected boolean shouldNotFilter(HttpServletRequest request) {
-        String path = request.getRequestURI();
-        return !path.startsWith("/api/v1/admin");
-    }
-
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
             throws ServletException, IOException {
 
         String token = resolveToken(request);
 
+        // 🌟 토큰이 있고, 타입이 "ADMIN"일 때만 어드민 인증 로직 수행
         if (token != null && adminJwtProvider.validateToken(token)) {
-            // [기존 코드] 정상적인 토큰일 경우 데이터 추출 및 인증
-            Long id = adminJwtProvider.getId(token);
-            String loginId = adminJwtProvider.getSubject(token);
-            String role = adminJwtProvider.getRole(token);
+            if ("ADMIN".equals(adminJwtProvider.getType(token))) {
+                Long id = adminJwtProvider.getId(token);
+                String loginId = adminJwtProvider.getSubject(token);
+                String role = adminJwtProvider.getRole(token);
 
-            CustomManagerDetails managerDetails = new CustomManagerDetails(id, loginId, role);
-            UsernamePasswordAuthenticationToken authentication = new UsernamePasswordAuthenticationToken(
-                    managerDetails, null, managerDetails.getAuthorities()
-            );
-            SecurityContextHolder.getContext().setAuthentication(authentication);
-
-        } else {
-            // 🚨 [테스트용 임시 코드] 토큰이 없거나 틀려도 어드민 API에서는 무조건 강제 통과!
-            System.out.println("⚠️ [TEST MODE] 인증 없이 무적 권한(SUPER_ADMIN)으로 강제 통과시킵니다.");
-
-            // PK 1번, 아이디 test_admin, 권한 SUPER_ADMIN을 가진 가짜 매니저 객체 생성
-            CustomManagerDetails dummyManager = new CustomManagerDetails(1L, "test_admin", "ROLE_SUPER_ADMIN");
-
-            UsernamePasswordAuthenticationToken dummyAuth = new UsernamePasswordAuthenticationToken(
-                    dummyManager, null, dummyManager.getAuthorities()
-            );
-            // 시큐리티에 가짜 신분증 제출
-            SecurityContextHolder.getContext().setAuthentication(dummyAuth);
+                CustomManagerDetails managerDetails = new CustomManagerDetails(id, loginId, role);
+                UsernamePasswordAuthenticationToken authentication = new UsernamePasswordAuthenticationToken(
+                        managerDetails, null, managerDetails.getAuthorities()
+                );
+                SecurityContextHolder.getContext().setAuthentication(authentication);
+            }
         }
+        // (테스트용 강제 통과 로직은 필요에 따라 유지/삭제)
 
         filterChain.doFilter(request, response);
+    }
+
+    // 이미 잘 작성되어 있는 부분 (유지)
+    @Override
+    protected boolean shouldNotFilter(HttpServletRequest request) {
+        String path = request.getRequestURI();
+        return !path.startsWith("/api/v1/admin");
     }
 
     private String resolveToken(HttpServletRequest request) {
