+++ src/main/java/com/kidmily/algoga_server/global/config/GlobalSecurityConfig.java
@@ -1,0 +1,16 @@
+package com.kidmily.algoga_server.global.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@Configuration
+public class GlobalSecurityConfig {
+
+    // 🌟 User와 Admin 도메인 모두에서 공통으로 사용할 PasswordEncoder
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+}
