--- src/main/java/com/kidmily/algoga_server/global/config/SecurityConfig.java
@@ -1,59 +1,0 @@
-package com.kidmily.algoga_server.global.config;
-
-import com.kidmily.algoga_server.admin.settings.AdminJwtAuthenticationFilter;
-import com.kidmily.algoga_server.user.settings.JwtAuthenticationFilter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-@Configuration
-@EnableWebSecurity
-@RequiredArgsConstructor
-public class SecurityConfig {
-
-    private final JwtAuthenticationFilter jwtAuthenticationFilter;
-    private final AdminJwtAuthenticationFilter adminJwtAuthenticationFilter; // 어드민 필터 주입
-
-    // PasswordEncoder를 한 곳에서 관리하기 위해 여기에 빈 등록
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http
-                .cors(cors -> {})
-                .csrf(AbstractHttpConfigurer::disable)
-                .formLogin(AbstractHttpConfigurer::disable)
-                .httpBasic(AbstractHttpConfigurer::disable)
-                .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-                .authorizeHttpRequests(auth -> auth
-//                        .requestMatchers(
-//                                "/api/v1/auth/**",
-//                                "/swagger-ui/**",
-//                                "/v3/api-docs/**",
-//                                "/api/v1/countries/**",
-//                                "/api/v1/packages/**",
-//                                "/api/v1/bookings/**",
-//                                "/api/v1/payments/**",
-//                                "/api/v1/refund-requests/**",       // 추가
-//                                "/api/v1/admin/refund-requests/**"  //테스트용 추가
-//                        ).permitAll()
-                                .anyRequest().permitAll()
-
-                )
-                .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class)
-                .addFilterAfter(adminJwtAuthenticationFilter, JwtAuthenticationFilter.class); // 어드민 필터 추가
-
-        return http.build();
-    }
-}
